+++ src/index.ts
@@ -1,5 +1,5 @@
 export { Client, ClientError } from "./client";
 export type { ClientOptions } from "./client";
-export { Server, writeSSEEvents } from "./server";
+export { Server } from "./server";
 export type { ServerHandler, ServerOptions } from "./server";
 export type * from "./types";